Day12
Controllers와 Routers를 추가하거나 제거해서는 안됩니다. 단 3가지만으로 아래 챌린지를 수행할 수 있어요.

getMovies returns an array of movies, console.log the result to see the shape of the object.

getMovieById returns a movie object. It requires an ID as an argument and if no movie is found it returns undefined.

getMovieByMinimumRating returns an array of movies, it requires a number, with this number the function will return an array of movies with a rating equal or higher than the number.

getMovieByMinimumYear returns an array of movies, it requires a number, with this number the function will return an array of movies with a release date equal or higher than the number.

You can only have 3 .pug pages. movies.pug, 404.pug, detail.pug

Use middlewares to add locals like siteTitle

Use pug and mixins to render pages and loop over the movies list.